copilot-theorem: Add functionality for bisimulation proofs of Copilot specs
In pursuit of proving the correctness of `copilot-c99`–generated code, this
patch introduces the `computeBisimulationProofBundle` function, which returns
all of the symbolic states necessary to carry out a bisimulation proof. The
`BisimulationProofBundle` that this returns facilitates performing
k-induction on all the properties in a given specification, where k is chosen
to be the maximum amount of delay on any of the involved streams.

In order to make this possible, a fair bit of refactoring had to be made to the
internals of `TransM` and the associated What4 translation machinery. While
related, the bisimulation machinery is doing something separate from the bits
that translate Copilot specifications directly to What4 formulas. For this
reason, this patch moves all of the translation-related functions to a new
`Copilot.Theorem.What4.Translate` module, which helps to keep this part of
`copilot-theorem` more self-contained.