Add special case handling for length-1 stream buffers.
In these cases, a ring buffer is unnecessary, as there is only one
element, so we omit the buffer index varible and use a single global
of the correct type instead of an array.

This simplifies this relatively common case, and prevents a
problem that was occuring sometimes in the verifier.
In particular, `clang` is clever enough to do this optimization on
it's own when compiling at `-O1` or higher (but not at `-O0`).
In those cases, the verifier was not finding the global values
where it expected (because clang had modified and renamed them).
Instead, we now do handle this special-case optimization ourselves
both in the code generator and in the verifier.

Fixes #258