Allow properties to be used as assumptions when setting up bisimuation proofs.
This is sometimes helpful when clang finds inductive properties of your program
and simplifies using them. This way, the user can (if they are clever) discover
what property is necessary to allow the proof to go through, prove it by
whatever means convenient, and then assume that property when doing program
verification.