Use `SymInterpretedFloat`s in `XExpr` (#1)
Previously, the `XFloat` and `XDouble` constructors of `XExpr` were using
IEEE-754 representations of floating-point numbers. However, this doesn't
mesh well with our current plans for `copilot-verifier`, in which we want
to use uninterpreted floats. To fix this, we instead of `SymInterpretedFloat`s
that are polymorphic over the particular What4 representation of floating-point
numbers. This involves a fair amount of downstream code churn, but for the
most part, this is mechanical.

This is one step towards GaloisInc/copilot-verifier#3.