Avoid catch-all wildcard patterns in translateOp{1,2,3}
Previously, the use of catch-all wildcard patterns made it appear as though we
had handled every case in `translateOp{1,2,3}`, but this was not actually the
case. This patch refactors things slightly to make sure that if we ever add
more `Op{1,2,3}`s in the future, then GHC's pattern-match coverage checker will
notify us that we need to add cases for them.