Make C.Compile.C99.CodeGen.typetypes return types in dependency order
Without this, Copilot programs involving nested structs had the potential to
omit struct declarations if their types weren't directly mentioned in a
trigger. The fix is simple—make `typetypes` recurse into struct types to make
sure that nested struct types are also handled. While I was in town, I also did
the same thing for array types. I don't have a program on hand involving arrays
that exhibits the same buggy behavior, but it seems best to be consistent in
our treatment of structs and arrays.

This is a workaround for Copilot-Language/copilot#275. This is needed in order
to verify examples involving nested structs—see GaloisInc/copilot-verifier#9.