Make struct arguments pointers in trigger functions
This is a design choice that I have adopted for the purpose of making the LLVM
bitcode that Clang generates easier to analyze. The issue is that if you have
a trigger function like this:

```c
void trigger(struct s ss) { ... }
```

The resulting LLVM bitcode will likely look very different from the source C
code, since LLVM does a fair bit of transformation to fit the fields of
`struct s` into a form that complies with the System V ABI. For more discussion
on this point, see GaloisInc/copilot-verifier#9.

We can bypass this issue entirely by making trigger functions pass structs by
reference, not by value. That is, do this:

```c
void trigger(const struct s *ss) { ... }
```

Clang compiles a C struct pointer directly to an LLVM struct pointer, which
makes it much easier to deal with in a consistent manner.