copilot-theorem: Split out translation functionality into its own module
The functionality for translating Copilot values into `what4` terms is
self-contained enough to warrant belonging in its own internal module, which we
have christened `Copilot.Theorem.What4.Translate`. In separate commits, we will
also add additional functionality to this module, so splitting it out will
prevent `Copilot.Theorem.What4` from becoming too bloated.

This patch does nothing but move the corresponding definitions to the new
module and rearrange the definitions slightly so that they are oriented more
from most important to least important. In particular, this does not make an
effort to make all of the code adhere to the Copilot style guide, as that will
occur in separate commits.

Co-authored-by: Rob Dockins <rdockins@galois.com>