copilot-theorem: Use what4's sym interface
Instead of directly relying on the `what4` `ExprBuilder` type, we now code
against the polymorphic `sym` interface, relying on the `IsExprBuilder` and
`IsSymExprBuilder` classes for performing various symbolic operations. This is
purely a refactoring, but it makes it more obvious which parts of the code are
using the various features that `what4` offers.

Co-authored-by: Rob Dockins <rdockins@galois.com>