copilot-theorem: Don't roundtrip through real-numbered operations
Previously, `copilot-theorem` assumed the use of IEEE floating-point numbers,
which meant that it had to hackily roundtrip through real-numbered operations in
order to implement Copilot operations such as `Logb` and `Pow`. Now that
`copilot-theorem` supports uninterpreted floats, however, there is no longer any
reason to do this, as we can just as well use `what4`'s built-in support for
special floating-point operations.

This patch replaces all of the various real-numbered operations with
`iFloatSpecialFunction{1,2}`. This allows us to delete a fair bit of unused
code, including:

* The `RealOp{1,2}` type synonyms
* The `logb` and `pow` fields of `TransState`

A consequence of the second bullet point above is that `copilot-theorem` no
longer needs to generate its own uninterpreted functions in `runTransM` to
represent Copilot's `Logb` and `Pow` operations. As a consequence, the
implementation of `runTransM` has been simplified, so the `sym` argument to
`runTransM` is no longer necessary.

Co-authored-by: Rob Dockins <rdockins@galois.com>