copilot-theorem: Make Sign translation match copilot-c99
`copilot-c99` recently changed the way it translates the `Sign` operation to C
code to make its semantics match those of Haskell's `signum` function
(see #278). This updates the corresponding translation of `Sign` in
`copilot-theorem` to use the same semantics. This is especially important for
the floating-point version of `Sign`, where the use of uninterpreted functions
requires that the order of operations closely match what is used in the C code.

Co-authored-by: Rob Dockins <rdockins@galois.com>