copilot-theorem: Track side conditions from partial function applications
Some Copilot operations can be partial, such as `Div`. We would like to track
the conditions under which these operations are well defined such we can check
if the corresponding partial operations in `copilot-c99`–generated code are also
well defined. For instance, `Div` is only well defined if the second argument
is not zero, so we record this as a `Pred` and track it in `TransM`'s state.
We track similar side conditions for things like arithmetic overflow and
out-out-bounds array indexing.

One thing we currently do not track are side conditions for floating-point
operations (e.g., `Fdiv`). See `Note [Side conditions for floating-point
operations]`. In case we wish to revisit this choice, I have added comments next
to each floating-point operation describing the circumstances under which they
can be partial.

Co-authored-by: Rob Dockins <rdockins@galois.com>