copilot-theorem: Simplify the handling of relative/absolute stream indexing
Previously, streams were indexed by `Int`s, which made it difficult to track
whether we were dealing with absolute offsets (i.e., offsets relative to the
beginning of time t=0) or relative offsets (i.e., offsets relative to some
arbitrary, but fixed j>=0). To make this easier to track, this patch removes the
`Int` indexing in favor of indexing by the new `StreamOffset` data type, which
clearly separates the `AbsoluteOffset` and `RelativeOffset` cases.

Some refactoring was done as a consequence of updating the relevant code to use
`StreamOffset`:

* The `translateExpr` and `translateExprAt` functions were combined, as the
  differences between the two functions are now encapsulated in `StreamOffset`.
* Similarly, the `getExternConstant` and `getExternConstantAt` functions have
  been combined.
* The `streamConstants` field of `TransState` has been renamed to
  `streamValues`, as this now stores more than just constants. Similarly, the
  `getStreamConstant` function has been renamed to `getStreamValue`.

Co-authored-by: Rob Dockins <rdockins@galois.com>