copilot-theorem: Use k-induction in Copilot.Theorem.What4.prove
Previously, `prove` did not perform any sort of induction when proving theorems.
However, it is possible to perform k-induction on the number of elements in the
buffers used in a Copilot program. The trick is to figure out what `k` should
be, as the number of base cases one needs in k-induction depends on how large
the buffers are. In this patch, we pick the heuristic of choosing `k` to be the
the maximum buffer length used in a Copilot program, which is often effective at
proving Copilot specifications.

Co-authored-by: Rob Dockins <rdockins@galois.com>