copilot-theorem: Support bisimulation proofs involving copilot-c99 code
While the existing `prove` function is suitable for proving properties about
Copilot specifications, they do not say anything about the C code that would be
generated from Copilot programs by `copilot-c99`. To that end, this patch
introduces functionality needed for bisimulation proofs between Copilot programs
and the corresponding C code. The following definitions have been introduced:

* `computeBisimulationProofBundle`, which computes all of the symbolic states
  necessary to carry out a bisimulation proof. These states are contained within
  the `BisimulationProofBundle` value that it returns.
* The `BisimulationProofBundle` data type, which bundles all of the necessary
  symbolic states for the proof. These include the initial state of the global
  variables, the states before and after the `step` function, the values of
  external streams, user-provided property assumptions, and any side conditions
  that arise from applications of partial operations in the Copilot program.
* The `BisimulationProofState` data type, which describes the state of a stream
  at some point in time in a Copilot program.

In addition, `Copilot.Theorem.What4` now re-exports the `XExpr` data type from
`Copilot.Theorem.What4.Translate`, as `XExpr` is used in the fields of
`BisimulationProofBundle` and `BisimulationProofState`.

The task of actually orchestrating the bisimulation proof is outside the scope
of the `copilot-theorem` package, but this patch exposes all of the necessary
pieces of functionality so that external tools can perform the proof.

Co-authored-by: Rob Dockins <rdockins@galois.com>