DRAFT: copilot-core: Use generic default implementations of Struct and Typed methods. Refs #???.
This updates the default implementations of the class methods of `Struct` and
`Typed` to use `Generic`-based default implementations. As such, if one derives
a `Generic` instance for a struct data type, it is now possible to define
`Struct` and `Typed` instances for the data type without needing to provide
explicit definitions for any of the methods. Alternatively, one can even derive
the `Struct` and `Typed` instances by using GHC's `DeriveAnyClass` language
extension.

This is a breaking change because any previously existing `Struct` instances
that omit an explicit definition of `updateField` and do not give the struct
data type a `Generic` instance will no longer compile. Migrating such instances
is relatively straightforward: simply derive a `Generic` instance for the
struct data type and it will compile once again.

This work is based off of an initial implementation by Marten Wijnja (@Qqwy)
from #561.

Co-authored-by: Marten Wijnja <marten.wijnja@channable.com>