DRAFT: copilot: Update struct examples to derive Struct and Typed instances. Refs #???.
The struct examples now use GHC's `DeriveAnyClass` language extension to derive
`Struct` and `Typed` instances for the data types used in the examples, greatly
reducing the amount of boilerplate code needed. Note that we continue to
manually implement the `updateField` method in the `Struct` instances in the
`StructsUpdateField` example, as this example is intended to demonstrate how
this is done for educational purposes.