copilot-theorem: Add XEmptyArray and XArray cases for valFromExpr. Refs #589.
Previously, the `valFromExpr` function was lacking cases for `XEmptyArray` and
`XArray`, so it would fail if the function was called on these values.

This commit adds these missing cases, which make use of the `Typed` evidence
added to `XEmptyArray` and `XArray` in a previous commit. We do not yet add a
case for structs, which prove more challenging.

This is an internal-only refactoring that should not have any changes in
user-facing behavior.