copilot-theorem: Remove unused CounterExample definition. Refs #589.
`CounterExample` was not exported from `Copilot.Theorem.What4`, and moreover,
its current definition was insufficient to capture all of the information that
would be desirable from a counterexample.

This commit removes `CounterExample` in preparation for a follow-up commit that
adds an improved version of `CounterExample`. As `CounterExample` was not
exported, this change is purely an internal refactoring.