copilot-theorem: Add Show and ShowF instances for CopilotValue. Refs #589.
Previously, the `CopilotValue` data type lacked a `Show` or `ShowF` instances,
which made it impractical to display them.

This commit adds a `Show` instance so that a value of type `CopilotValue a` can
be shown, and it also adds a `ShowF` instance so that a value of type `Some
CopilotValue` can be shown.