copilot-theorem: Add prove'. Refs #589.
Currently, the `Copilot.Theorem.What4.prove` function returns a list of
results, where each result contains a `SatResult` that describes whether a
property is `Valid`, `Invalid`, or `Unknown`. The `Invalid` result has the
limitation that it does not give any information about a specific
counterexample that could drive Copilot into falsifying the property, however.
This makes it challenging to interpret what the results of prove mean.

This introduces a new `prove'` function to `Copilot.Theorem.What4` that mirrors
the type signature of `prove`, except that it returns a variant of `SatResult`
where the `Invalid` equivalent encodes counterexample information.
`copilot-theorem` users can then interpret the results of the counterexample in
Copilot specifications.