Proposed fix for Copilot-Language/copilot#673
Adds a flag to enable testsuite if bsc installed (defaults to disabled). Better
solutions might be (1) split the "bsc" tests out into a separate testsuite so
more basic unit/regression tests can still be run without bsc or (2) not create
a separate (optionally buildable) testsuite but rather add a runtime flag to
the existing testsuite to enable the use of bsc.