added a fix for incorrectly converted special floating points (discussion/issue #675)
Previously copilot-bluespec used `realToFrac` to cast doubles to Float but this ignores special cases of +/-Inf, NaN, and -0.0.
This patch adds in those cases specifically.

the What4 usage in copilot-theorem has a similar use of `realToFrac` but it doesn't seem to be an issue.
if it does become an issue, an identical fix should work there.